Fix for Mautic\CampaignBundle\EventListener\StatsSubscriber::__construct() does not call parent constructor from Mautic\CoreBundle\EventListener\StatsSubscriber.
