+++ plugins/MauticFocusBundle/EventListener/StatsSubscriber.php
@@ -12,7 +12,7 @@
 namespace MauticPlugin\MauticFocusBundle\EventListener;
 
 use Doctrine\ORM\EntityManager;
-use Mautic\CoreBundle\EventListener\StatsSubscriber as CommonStatsSubscriber;
+use Mautic\CoreBundle\EventListener\CommonStatsSubscriber;
 
 /**
  * Class StatsSubscriber.
